+++ src/routes/SwitchDepartment/container.js
@@ -27,9 +27,8 @@ export class SwitchDepartmentContainer extends Component {
 		}
 
 		if (!room) {
-			await Livechat.grantVisitor({ visitor: { department, token } });
-			await loadConfig();
-			await dispatch({ alerts: (alerts.push({ id: createToken(), children: I18n.t('Department switched'), success: true }), alerts) });
+			const user = await Livechat.grantVisitor({ visitor: { department, token } });
+			await dispatch({ user, alerts: (alerts.push({ id: createToken(), children: I18n.t('Department switched'), success: true }), alerts) });
 			return history.go(-1);
 		}