+++ src/widget.js
@@ -231,6 +231,53 @@ function pageVisited(change) {
 	});
 }
 
+function initialize(params) {
+	for (const method in params) {
+		switch(method) {
+			case 'customField':
+				const { key, value, overwrite } = params[method];
+				setCustomField(key, value, overwrite);
+				continue;
+			case 'setCustomFields':
+				if (!Array.isArray(params[method])) {
+					console.log('Error: Invalid parameters. Value must be an array of objects');
+					continue;
+				}
+				params[method].forEach((data) => {
+					const { key, value, overwrite } = data;
+					setCustomField(key, value, overwrite);
+				});
+				continue;
+			case 'theme':
+				setTheme(params[method]);
+				continue;
+			case 'department':
+				setDepartment(params[method]);
+				continue;
+			case 'guestToken':
+				setGuestToken(params[method]);
+				continue;
+			case 'guestName':
+				setGuestName(params[method]);
+				continue;
+			case 'guestEmail':
+				setGuestEmail(params[method]);
+				continue;
+			case 'registerGuest':
+				registerGuest(params[method]);
+				continue;
+			case 'language':
+				setLanguage(params[method]);
+				continue;
+			case 'agent':
+				setAgent(params[method]);
+				continue;
+			default:
+				continue;
+		}
+	}
+}
+
 function setCustomField(key, value, overwrite) {
 	if (typeof overwrite === 'undefined') {
 		overwrite = true;
@@ -264,6 +311,10 @@ function registerGuest(guest) {
 
 function clearDepartment() {
 	callHook('clearDepartment');
+}
+
+function setAgent(agent) {
+	callHook('setAgent', agent);
 }
 
 function setLanguage(language) {
@@ -350,12 +401,14 @@ window.RocketChat.livechat = {
 	// methods
 	pageVisited,
 	setCustomField,
+	initialize,
 	setTheme,
 	setDepartment,
 	clearDepartment,
 	setGuestToken,
 	setGuestName,
 	setGuestEmail,
+	setAgent,
 	registerGuest,
 	setLanguage,
 	showWidget,